fix fee estimation bug in autoloop
When we query the quote from the server in autoloop, before this commit
it uses an on-chain fees estimated by them, This will be imcompatible when
actually executing the swap. Because we use a feerate estimated by our
FeeEstimator. This commit fixes it by using FeeEstimator also when getting
loopin/out quotes.